python-service/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.1
langchain>=0.1.11
langchain-community>=0.0.27
langchain-core>=0.1.30
python-multipart>=0.0.9
pypdf>=4.1.0
python-docx>=1.1.0
docx2txt>=0.8
faiss-cpu>=1.8.0
sentence-transformers>=2.5.1
dashscope>=1.14.0
python-dotenv>=1.0.1
pymilvus>=2.4.0
langchain-milvus>=0.1.0
pillow>=10.3.0
pytesseract>=0.3.10
